requirement.txt
absl-py==0.9.0
appdirs==1.4.3
appnope==0.1.0
astunparse==1.6.3
attrs==19.3.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
black==19.10b0
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
entrypoints==0.3
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.9
ipykernel==5.3.0
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1590796925971/work
ipython-genutils==0.2.0
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1592619347183/work
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib @ file:///opt/concourse/worker/volumes/live/59859a4b-794b-4c9f-51fd-a0ac1d1ab595/volume/matplotlib-base_1592846058943/work
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mypy-extensions==0.4.3
nb-black==1.0.7
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.2.1
pandas @ file:///opt/concourse/worker/volumes/live/5ef6147e-ea92-446e-4065-0c8c7435293f/volume/pandas_1592841676717/work
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1592500439797/work
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pyzmq==18.1.1
regex @ file:///Users/runner/miniforge3/conda-bld/regex_1591646146133/work
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
seaborn==0.11.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1589469402899/work
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1588470653596/work
urllib3==1.25.9
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1592931742287/work
Werkzeug==1.0.1
wrapt==1.12.1
